Add sortingTom.go file and update channel.go documentation
A new file, sortingTom.go, is created to define a new 'element' type and use it in a slice. This commit also includes an update to the documentation in channel.go, improving the clarity of the comments regarding potential deadlocks and the use of mutex locks with channels. Moreover, a new go.mod file has been added specifying the Go version used and the module name.