Add error handling to user input in calculator
The previous version of the calculator would crash if the user input was not expected. This update includes importing and using the 'log' package, and implementing string scan error checking. Now, if there is an error while scanning the user input, a log message will be printed describing the error before a fatal exit.