instructions in README
The README instructions have been updated to include steps on how to create a .exe file for the project. Preexisting dependencies on Golang installation are now specifically called out. Further steps might be needed for full usability on all systems, as the ability to directly include the .exe file in this repository is still being explored.