Refactor chess board layout and implement grid system
This update includes significant changes to the structure of the chess board in the 'chess.go' file. We have included a gridCoordinate array into the square struct to utilize a grid system representation. Additionally, adjustments have been made to the printBoard and createBoard functions to accommodate these modifications. We've also made a minor change in the 'channel.go' file to alter the output format from Println to Print.