Refactor code to check kings' status instead of queens'
Changed the functionality from checking if the queens are still in play to checking if the kings are still in play in a chess game. The changes affect how the game determines the player in winning position. This change also implies that, instead of moving a piece after looking for the queens' status, the program will first move the piece then look for the kings' status.